requirements.txt
fastapi
uvicorn
sse-starlette
pydantic
langgraph
langchain-google-genai
langchain-core
python-dotenv
wikipedia
duckduckgo-search
ddgs
